requirements.txt
Flask==3.1.1
Werkzeug==3.1.3

torch==2.6.0
transformers==4.52.3
tokenizers==0.21.1
safetensors==0.5.3

pandas==2.3.0
numpy==1.26.4
scikit-learn==1.4.2

langdetect==1.0.9
nltk==3.9.1
ekphrasis==0.5.4
chardet==5.2.0
beautifulsoup4==4.12.3

mysql-connector-python==9.3.0
plotly==6.1.2
wordcloud==1.9.4
tqdm==4.67.1
